Enhancements
Better context for HTML elements. HTML is now returned by block level elements, and the elements selector is given as context. Attributes also return a selector as context and are returned individually. Html comments are returned as individual hunks.
Add Stylesheet and cpp filters. JavaScript is now derived from cpp. PySpelling looks for .spelling.yml or .pyspelling.yml with a priority for the latter.