Wave 2 enhancements
Create a util function for deprecation warnings.  Move wildcard flow with its own named pluig separating it from the reference base class. Provide a class attribute that will cause a Filter object to avoid BOM detection if it is not appropriate for the given file. If the first entry in a pipeline is a FlowControl plugin, complain. Wordlists should get the desired language/dictionary from the spell checker specific options. Improve main page documentation.