Improvments for ooxml and odf
Add reset method to all plugins and call on run.  Soft break on paragraphs, but only do hard breaks on presentation slides and provide context on which slide.  Don't display XML selector as it isn't helpful in document context. Read mimetypes of document files to determine what type of file we are dealing with and adust accordingly. No need to specify a mode. Add extract_tag_metadata to xml class so we can capture tag data during the recursion.