Rework comments in XML (#71)
* Rework comments in XML
There is technically nothing wrong with what we are doing, but the
unreleased Soup Sieve 1.9 will be deprecating the comment extraction
API and will focus solely on CSS selectors moving forward. So we will
go ahead and implement comment extraction ourselves.

Also update copyright and document CSS.

* Fix tests
Newer soupsieve will not parse empty patterns and will not treat
elements that do not have namespace as having a namespace simply because
a colon is the name of the attribute.