Prepare for wcmatch 5.0
I will be releasing wcmatch 5.0 which removes an old flag FORCECASE.
It is that useful as it's behavior would force Unix style path logic if
used on a Windows system. For glob, which accesses the filesystem, it
was ignored on Windows anyways. Basically since Unix/Linux behavior is
case sensitive by default, it was a pointless flag to expose in our
environment.

It has been replaced with CASE which actually is more useful and allows
for forcing case sensitivity on a Windows system. Still probably won't
get any use, but we need to account for this change so not break.