Add support for wcmatch 6
wcmatch 6 will ensure that only unique results get returned. This is
probably the most useful feature that is gained from wcmatch 6.

It also imposes a limit of 1000 patterns in one call. Mainly this
is to protect against pattern expansions, such as with braces:
   {1..10000000000}
1000 should be plenty, but we expose the ability to change this value
or disable it. As this project is mainly used in CI environments where
the maintainer controls this, it is very low risk for someone to
maliciously inject big patterns anyways, but moving forward, we will
support controlling this.

We also expose tilde expansion, which really isn't something that people
should be using in this type of project, but for completeness, it is
exposed.