Remove tests from installation
setuptools requires you provide a pattern that prevents the top level
from installing plus the sub modules under the top level. One would
think if the top level is excluded that the children would also be
excluded, and yet this isn't the case. I guess this is to allow fine
control for people including things in weird ways. This fixes the
issue by provided a wildcard pattern that excludes both top and subs.